move factory classes to root directory
