--- Makefile
@@ -1,2 +1,0 @@
-run_tests:
-    python -m unittest discover
