--- ___setup.cfg
@@ -1,28 +1,0 @@
-# Inside of setup.cfg
-[metadata]
-name = pytfsclient
-version = 1.2
-author = TopTuK
-author_email = cydoor88@gmail.com
-description = Python Microsoft Team Foundation Server Library is a  client that can work with Microsoft TFS workitems
-long_description = file: README.md
-long_description_content_type = text/markdown
-url = https://github.com/TopTuK/PyTfsClient
-project_urls =
-    Bug Tracker = https://github.com/TopTuK/PyTfsClient/issues
-classifiers =
-    Development Status :: 4 - Beta
-    Intended Audience :: Developers
-    Topic :: Software Development :: Build Tools
-    License :: OSI Approved :: MIT License
-    Programming Language :: Python :: 3.8
-
-[options]
-package_dir =
-    = src
-#packages = find:
-
-python_requires = >=3.6
-
-[options.packages.find]
-where = src
