+++ src/pytfsclient/services/project_client/project_client.py
@@ -3,7 +3,7 @@ from ...models.project.tfs_project import TfsProject
 from ...models.project.tfs_team import TfsTeam
 from ...models.project.tfs_team_member import TfsTeamMember
 from ...services.base_client import BaseClient
-from ...services.client_connection import ClientConnection
+from ...client_connection import ClientConnection
 from typing import List
 
 class ProjectClient(BaseClient):
